+++ spec/osm/badge_spec.rb
@@ -95,9 +95,19 @@ describe "Badge" do
 
   it "Get total requirements gained for a member" do
     data = Osm::Badge::Data.new(
-      :requirements => {'a_1' => 'x', 'a_2' => 'a', 'b_1' => 'yes', 'b_2' => '2000-01-02'}
+      :requirements => {
+        'a_1' => 'a',
+        'a_2' => 'yes',
+        'a_3' => '2000-01-02',
+        'a_4' => 1,
+        'b_1' => 'x',
+        'b_2' => 'xYES',
+        'b_3' => '',
+        'b_4' => nil,
+        'b_5' => 0,
+      }
     )
-    data.total_gained.should == 3
+    data.total_gained.should == 4
   end
 
   it "Get total requirements met in each section for a member" do
@@ -108,13 +118,21 @@ describe "Badge" do
         Osm::Badge::Requirement.new(:field => 'a_2'),
         Osm::Badge::Requirement.new(:field => 'b_1'),
         Osm::Badge::Requirement.new(:field => 'b_2'),
+        Osm::Badge::Requirement.new(:field => 'y_1'),
+        Osm::Badge::Requirement.new(:field => 'y_2'),
       ]
     )
     data = Osm::Badge::Data.new(
       :badge => badge,
-      :requirements => {'a_1' => 'x', 'a_2' => '', 'b_1' => 'yes', 'b_2' => '2000-01-02'}
+      :requirements => {
+        'a_1' => 'x',
+        'a_2' => '',
+        'b_1' => 'yes',
+        'b_2' => '2000-01-02',
+        'y_1' => 1,
+      }
     )
-    data.gained_in_sections.should == {'a' => 0, 'b' => 2}
+    data.gained_in_sections.should == {'a' => 0, 'b' => 2, 'y' => 1}
   end
 
   it "Get number of sections met for a member" do
@@ -135,11 +153,105 @@ describe "Badge" do
     data.sections_gained.should == 2
   end
 
-  it "Works out if the badge is due" do
-    Osm::Badge::Data.new(:completed => 0, :awarded => 0).due?.should be_false
-    Osm::Badge::Data.new(:completed => 1, :awarded => 0).due?.should be_true
-    Osm::Badge::Data.new(:completed => 2, :awarded => 2).due?.should be_false
-    Osm::Badge::Data.new(:completed => 2, :awarded => 1).due?.should be_true
+  describe "Works out if the badge has been earnt" do
+    it "Staged" do
+      badge = Osm::StagedBadge.new(:osm_key => 'not_hikes_or_nights')
+      data = Osm::Badge::Data.new(:awarded => 2, :badge => badge)
+
+      data.stub(:earnt) { 1 }
+      data.earnt?.should be_false
+
+      data.stub(:earnt) { 2 }
+      data.earnt?.should be_false
+
+      data.stub(:earnt) { 3 }
+      data.earnt?.should be_true
+    end
+
+    it "Non staged" do
+      badge = Osm::ActivityBadge.new()
+      data = Osm::Badge::Data.new(:completed => 1, :awarded => 1, :badge => badge)
+      data.earnt?.should be_false
+
+      badge = Osm::ActivityBadge.new()
+      data = Osm::Badge::Data.new(:completed => 1, :awarded => 0, :badge => badge)
+      data.earnt?.should be_true
+
+
+      badge = Osm::ActivityBadge.new(:total_needed => 0, :sections_needed => 2, :needed_from_section => {'a' => 2, 'b' => 1})
+      data = Osm::Badge::Data.new(:requirements => {'a_01'=>'y', 'a_02'=>'y', 'b_01' => 'y'}, :completed => 0, :awarded => 0, :badge => badge)
+      data.earnt?.should be_true
+
+      badge = Osm::ActivityBadge.new(:total_needed => 0, :sections_needed => 2, :needed_from_section => {'a' => 2, 'b' => 1})
+      data = Osm::Badge::Data.new(:requirements => {'a_01'=>'y', 'a_02'=>'y', 'b_01' => 'x'}, :completed => 0, :awarded => 0, :badge => badge)
+      data.earnt?.should be_false
+
+      badge = Osm::ActivityBadge.new(:total_needed => 3, :sections_needed => 0, :needed_from_section => {'a' => 2, 'b' => 1})
+      data = Osm::Badge::Data.new(:requirements => {'a_01'=>'y', 'a_02'=>'y', 'b_01' => 'y'}, :completed => 0, :awarded => 0, :badge => badge)
+      data.earnt?.should be_true
+
+      badge = Osm::ActivityBadge.new(:total_needed => 3, :sections_needed => 0, :needed_from_section => {'a' => 2, 'b' => 1})
+      data = Osm::Badge::Data.new(:requirements => {'a_01'=>'y', 'a_02'=>'x', 'b_01' => 'y'}, :completed => 0, :awarded => 0, :badge => badge)
+      data.earnt?.should be_false
+
+      badge = Osm::ActivityBadge.new(:total_needed => 3, :sections_needed => 2, :needed_from_section => {'a' => 2, 'b' => 1})
+      data = Osm::Badge::Data.new(:requirements => {'a_01'=>'y', 'a_02'=>'y', 'b_01' => 'y'}, :completed => 0, :awarded => 0, :badge => badge)
+      data.earnt?.should be_true
+
+      badge = Osm::ActivityBadge.new(:total_needed => 1, :sections_needed => 1, :needed_from_section => {'a' => 2, 'b' => 1})
+      data = Osm::Badge::Data.new(:requirements => {'a_01'=>'y', 'a_02'=>'y', 'b_01' => 'y'}, :completed => 0, :awarded => 0, :badge => badge)
+      data.earnt?.should be_true
+    end
+  end
+
+  describe "Works out what level of a badge has been earnt" do
+    it "Staged" do
+      badge = Osm::StagedBadge.new(:osm_key => 'not_hikes_or_nights', :needed_from_section => {'a'=>1,'b'=>1,'c'=>1,'d'=>2,'e'=>2})
+
+      data = Osm::Badge::Data.new(:requirements=>{'a_01'=>'','b_01'=>'','c_01'=>'','d_01'=>'','d_02'=>'','e_01'=>'','e_02'=>''}, :badge=>badge)
+      data.earnt.should == 0
+
+      data = Osm::Badge::Data.new(:requirements=>{'a_01'=>'y','b_01'=>'','c_01'=>'','d_01'=>'','d_02'=>'','e_01'=>'','e_02'=>''}, :badge=>badge)
+      data.earnt.should == 1
+
+      data = Osm::Badge::Data.new(:requirements=>{'a_01'=>'y','b_01'=>'y','c_01'=>'','d_01'=>'y','d_02'=>'','e_01'=>'','e_02'=>''}, :badge=>badge)
+      data.earnt.should == 2
+
+      data = Osm::Badge::Data.new(:requirements=>{'a_01'=>'y','b_01'=>'y','c_01'=>'','d_01'=>'y','d_02'=>'y','e_01'=>'','e_02'=>''}, :badge=>badge)
+      data.earnt.should == 4
+    end
+
+    it "Nights away" do
+      badge = Osm::StagedBadge.new(:osm_key => 'nightsaway')
+
+      Osm::Badge::Data.new(:requirements => {'y_01'=>3}, :badge => badge).earnt.should == 1
+      Osm::Badge::Data.new(:requirements => {'y_01'=>5}, :badge => badge).earnt.should == 5
+      Osm::Badge::Data.new(:requirements => {'y_01'=>6}, :badge => badge).earnt.should == 5
+      Osm::Badge::Data.new(:requirements => {'y_01'=>199}, :badge => badge).earnt.should == 175
+      Osm::Badge::Data.new(:requirements => {'y_01'=>200}, :badge => badge).earnt.should == 200
+      Osm::Badge::Data.new(:requirements => {'y_01'=>999}, :badge => badge).earnt.should == 200
+    end
+
+    it "Hikes away" do
+      badge = Osm::StagedBadge.new(:osm_key => 'hikes')
+
+      Osm::Badge::Data.new(:requirements => {'y_01'=>3}, :badge => badge).earnt.should == 1
+      Osm::Badge::Data.new(:requirements => {'y_01'=>5}, :badge => badge).earnt.should == 5
+      Osm::Badge::Data.new(:requirements => {'y_01'=>6}, :badge => badge).earnt.should == 5
+      Osm::Badge::Data.new(:requirements => {'y_01'=>49}, :badge => badge).earnt.should == 35
+      Osm::Badge::Data.new(:requirements => {'y_01'=>50}, :badge => badge).earnt.should == 50
+      Osm::Badge::Data.new(:requirements => {'y_01'=>999}, :badge => badge).earnt.should == 50
+    end
+
+    it "Non staged" do
+      data = Osm::Badge::Data.new(:badge => Osm::ActivityBadge.new)
+
+      data.stub(:earnt?) { true }
+      data.earnt.should == 1
+
+      data.stub(:earnt?) { false }
+      data.earnt.should == 0
+    end
   end
 
   it "Works out if the badge has been started" do